client/requirements.txt
grpcio
grpcio-tools
flask
flask-wtf
flask-cors